refactor(weave): move session SDK from trace/ to session/ package
Session SDK is a user-facing domain concept (like evaluation/, agent/,
prompt/), not tracing infrastructure. Moving it to its own top-level
package also avoids the circular import between session.py and
session_otel.py that would arise when OTel wiring lands.
